+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added field `active` for scenario schedulers. The field is required for POST/PUT endpoints, and it is also returned by GET endpoints.
+- Added endpoints `PUT /api/scenario-schedulers/:scenarioSchedulerId/activate` and `PUT /api/scenario-schedulers/:scenarioSchedulerId/activate`.
 
 ## 0.4.3 - 2023-07-10
 ### Changed